requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
googlemaps==4.10.0
ortools==9.10.4067
pydantic==2.9.2
python-dotenv==1.0.1
